common: Move the web server code into the common directory
So it can also be used in the bridge.
