+++ src/common/Makefile-common.am
@@ -55,6 +55,10 @@ libcockpit_common_a_SOURCES = \
 	src/common/cockpitunixfd.h \
 	src/common/cockpitunixsignal.c \
 	src/common/cockpitunixsignal.h \
+	src/common/cockpitwebresponse.h \
+	src/common/cockpitwebresponse.c \
+	src/common/cockpitwebserver.h \
+	src/common/cockpitwebserver.c \
 	$(NULL)
 
 libcockpit_common_a_CFLAGS = \
@@ -64,6 +68,7 @@ libcockpit_common_a_CFLAGS = \
 
 libcockpit_common_a_LIBS = \
 	libcockpit-common.a \
+	libwebsocket.a \
 	$(COCKPIT_LIBS) \
 	$(NULL)
 
@@ -77,6 +82,8 @@ COCKPIT_CHECKS = \
 	test-transport \
 	test-unixsignal \
 	test-template \
+	test-webresponse \
+	test-webserver \
 	$(NULL)
 
 test_hex_CFLAGS = $(libcockpit_common_a_CFLAGS)
@@ -102,6 +109,17 @@ test_transport_LDADD = $(libcockpit_common_a_LIBS)
 test_unixsignal_CFLAGS = $(libcockpit_common_a_CFLAGS)
 test_unixsignal_SOURCES = src/common/test-unixsignal.c
 test_unixsignal_LDADD = $(libcockpit_common_a_LIBS)
+
+test_webresponse_SOURCES = \
+	src/common/test-webresponse.c \
+	src/common/mock-io-stream.c src/common/mock-io-stream.h \
+	$(NULL)
+test_webresponse_CFLAGS = $(libcockpit_common_a_CFLAGS)
+test_webresponse_LDADD = $(libcockpit_common_a_LIBS)
+
+test_webserver_SOURCES = src/common/test-webserver.c
+test_webserver_CFLAGS = $(libcockpit_common_a_CFLAGS)
+test_webserver_LDADD = $(libcockpit_common_a_LIBS)
 
 noinst_PROGRAMS += $(COCKPIT_CHECKS)
 TESTS += $(COCKPIT_CHECKS)
