+++ src/common/cockpitwebresponse.c
