+++ src/common/cockpitwebresponse.h
