+++ src/common/mock-io-stream.c
