+++ src/common/mock-io-stream.h
