--- src/ws/cockpitwstypes.h
@@ -1,29 +1,0 @@
-/*
- * This file is part of Cockpit.
- *
- * Copyright (C) 2013 Red Hat, Inc.
- *
- * Cockpit is free software; you can redistribute it and/or modify it
- * under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation; either version 2.1 of the License, or
- * (at your option) any later version.
- *
- * Cockpit is distributed in the hope that it will be useful, but
- * WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with Cockpit; If not, see <http://www.gnu.org/licenses/>.
- */
-
-#ifndef __COCKPIT_BRIDGE_TYPES_H__
-#define __COCKPIT_BRIDGE_TYPES_H__
-
-#include <glib-unix.h>
-#include <gio/gio.h>
-
-struct _CockpitWebServer;
-typedef struct _CockpitWebServer CockpitWebServer;
-
-#endif /* __COCKPIT_BRIDGE_TYPES_H__ */
